chore: created first screen with actions study: SwiftUI + MVVM simple example
