+++ StudyProject-Swift/StudyProject-Swift/Study/Architectures/SwiftUI+MVVM/Models/User.swift
@@ -1,0 +1,11 @@
+//
+//  User.swift
+//  StudyProject-Swift
+//
+//  Created by Denis Svichkarev on 21/11/24.
+//
+
+struct User {
+    var name: String
+    var age: Int
+}
